moved textfields to Form component & added useEffects
