+++ src/pages/DiaryHome.jsx
@@ -1,35 +1,49 @@
-import Box from '@mui/material/Box';
-import Button from '@mui/material/Button';
-import FormControl from '@mui/material/FormControl';
-import Grid from '@mui/material/Grid';
-import InputLabel from '@mui/material/InputLabel';
-import OutlinedInput from '@mui/material/OutlinedInput';
-import DiaryCard from '../components/DiaryCard.jsx';
-import Title from '../components/Title.jsx';
-import MainHeader from '../components/MainHeader.jsx';
-import { purple } from '@mui/material/colors';
-import { styled } from '@mui/material/styles';
-import { addDoc, doc, updateDoc } from 'firebase/firestore';
-import { useState } from 'react';
-import { useSelector } from 'react-redux';
-import colRef, { db } from '../firebaseConfig';
+import { addDoc, doc, onSnapshot, updateDoc } from 'firebase/firestore';
+import { useEffect, useState } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
+import DiaryCard from '../components/DiaryCard.jsx';
+import Form from '../components/Form';
+import MainHeader from '../components/MainHeader.jsx';
+import Title from '../components/Title.jsx';
+import colRef, { db, totalDiariesRef } from '../firebaseConfig';
+import { diaryActions } from '../store/diary-slice';
 
 const DiaryHome = () => {
   const params = useParams();
+  const dispatch = useDispatch();
   const totalDiaries = useSelector((state) => state.diary.totalDiaries);
   const [title, setTitle] = useState('');
   const [description, setDescription] = useState('');
   const [cardColor, setCardColour] = useState('');
 
-  const SubmitButton = styled(Button)(({ theme }) => ({
-    color: theme.palette.getContrastText(purple[500]),
-    backgroundColor: '99c1ff',
-    ml: 1,
-    '&:hover': {
-      backgroundColor: '88b6ff',
-    },
-  }));
+  useEffect(() => {
+    dispatch(diaryActions.getDiaries());
+  }, [dispatch]);
+
+  useEffect(() => {
+    const unsubscribeDiaries = onSnapshot(colRef, (snapshots) => {
+      let diaries = [];
+      snapshots.docs.map((doc) =>
+        diaries.push({
+          id: doc.id,
+          ...doc.data(),
+        })
+      );
+      dispatch(diaryActions.replaceDiary(diaries));
+    });
+
+    const unsubscribeTotalDiaries = onSnapshot(totalDiariesRef, (snapshots) => {
+      let totalDiaries;
+      snapshots.docs.map((doc) => (totalDiaries = doc.data().totalDiaries));
+      dispatch(diaryActions.saveTotalDiaries(totalDiaries));
+    });
+
+    return () => {
+      unsubscribeDiaries();
+      unsubscribeTotalDiaries();
+    };
+  }, [dispatch]);
 
   const titleInputHandler = (event) => {
     event.preventDefault();
@@ -72,73 +86,16 @@ const DiaryHome = () => {
   return (
     <>
       <MainHeader />
-
       <Title />
-
-      <Grid container sx={{ alignItems: 'center', my: 2 }}>
-        <Grid item md={8} xs={8} sx={{ my: 1, display: 'flex' }}>
-          <Grid item md={8}>
-            <FormControl>
-              <InputLabel htmlFor='title'>Title</InputLabel>
-              <OutlinedInput
-                value={title}
-                type='text'
-                onChange={titleInputHandler}
-                id='title'
-                label='Title'
-                sx={{ borderRadius: 5 }}
-                placeholder='Title'
-                required
-              />
-            </FormControl>
-          </Grid>
-          <Grid item md={12} sx={{ ml: 1 }}>
-            <FormControl>
-              <InputLabel htmlFor='cardColor'>Card Colour</InputLabel>
-              <OutlinedInput
-                type='text'
-                value={cardColor}
-                onChange={cardColourInputHandler}
-                id='cardColor'
-                label='Card Color'
-                sx={{ borderRadius: 5 }}
-                placeholder='Hexadecimal color'
-                required
-              />
-            </FormControl>
-          </Grid>
-        </Grid>
-
-        <Grid item md={4} xs={4} sx={{ display: 'flex' }}>
-          <Box sx={{ flexGrow: 1 }} />
-          <SubmitButton
-            onClick={submitHandler}
-            size='large'
-            variant='contained'
-            sx={{ ml: 4, borderRadius: 5 }}
-          >
-            Submit
-          </SubmitButton>
-        </Grid>
-
-        <Grid item md={12} sx={{ my: 1 }} xs={12}>
-          <FormControl fullWidth sx={{ borderRadius: 5 }}>
-            <InputLabel htmlFor='description'>Description</InputLabel>
-            <OutlinedInput
-              value={description}
-              type='text'
-              onChange={descriptionInputHandler}
-              multiline
-              sx={{ borderRadius: 4 }}
-              rows={5}
-              id='description'
-              label='Description'
-              placeholder='Description'
-              required
-            />
-          </FormControl>
-        </Grid>
-      </Grid>
+      <Form
+        titleInputHandler={titleInputHandler}
+        descriptionInputHandler={descriptionInputHandler}
+        cardColourInputHandler={cardColourInputHandler}
+        submitHandler={submitHandler}
+        title={title}
+        cardColor={cardColor}
+        description={description}
+      />
       <DiaryCard />
     </>
   );
